+++ spec/rubocop/cop/rails/validation_spec.rb
@@ -58,6 +58,58 @@ RSpec.describe RuboCop::Cop::Rails::Validation do
 
         include_examples 'auto-corrects'
       end
+
+      context "with validates_#{type}_of when " \
+              'attributes are specified with array literal' do
+        let(:auto_corrected_source) do
+          "validates :full_name, :birth_date, #{type}: true"
+        end
+
+        let(:source) do
+          "validates_#{type}_of [:full_name, :birth_date]"
+        end
+
+        include_examples 'auto-corrects'
+      end
+
+      context "with validates_#{type}_of when " \
+              'attributes are specified with frozen array literal' do
+        let(:auto_corrected_source) do
+          "validates :full_name, :birth_date, #{type}: true"
+        end
+
+        let(:source) do
+          "validates_#{type}_of [:full_name, :birth_date].freeze"
+        end
+
+        include_examples 'auto-corrects'
+      end
+
+      context "with validates_#{type}_of when " \
+              'attributes are specified with symbol array literal' do
+        let(:auto_corrected_source) do
+          "validates :full_name, :birth_date, #{type}: true"
+        end
+
+        let(:source) do
+          "validates_#{type}_of %i[full_name birth_date]"
+        end
+
+        include_examples 'auto-corrects'
+      end
+
+      context "with validates_#{type}_of when " \
+              'attributes are specified with frozen symbol array literal' do
+        let(:auto_corrected_source) do
+          "validates :full_name, :birth_date, #{type}: true"
+        end
+
+        let(:source) do
+          "validates_#{type}_of %i[full_name birth_date].freeze"
+        end
+
+        include_examples 'auto-corrects'
+      end
     end
 
     context 'with single attribute name' do